chore: add environment types changeset (#176)
