+++ .changeset/add-environment-types.md
@@ -1,0 +1,6 @@
+---
+'@lapidist/design-lint': minor
+---
+
+add environment and token provider interfaces
+
