+++ src/core/index.ts
@@ -1,7 +1,13 @@
 export { Linter, type Config } from './linter.js';
 export { applyFixes } from './apply-fixes.js';
 export { Runner } from './runner.js';
-export type { DocumentSource, LintDocument } from './document-source.js';
+export type {
+  Environment,
+  TokenProvider,
+  SourceAdapter,
+  DocumentSource,
+  LintDocument,
+} from './environment.js';
 export type { PluginLoader, LoadedPlugin } from './plugin-loader.js';
 export type { CacheProvider, CacheEntry } from './cache-provider.js';
 export type {
