+++ src/node/file-document.ts
@@ -1,6 +1,6 @@
 import { readFile } from 'node:fs/promises';
 import path from 'node:path';
-import type { LintDocument } from '../core/document-source.js';
+import type { LintDocument } from '../core/environment.js';
 
 export function createFileDocument(filePath: string): LintDocument {
   const absPath = path.resolve(filePath);
