+++ src/node/file-source.ts
@@ -3,7 +3,7 @@ import { performance } from 'node:perf_hooks';
 import { realpathIfExists } from '../utils/paths.js';
 import { getIgnorePatterns } from '../core/ignore.js';
 import type { Config } from '../core/linter.js';
-import type { DocumentSource } from '../core/document-source.js';
+import type { DocumentSource } from '../core/environment.js';
 import { createFileDocument } from './file-document.js';
 
 const defaultPatterns = [
